+++ src/Microsoft.SqlTools.ServiceLayer/SchemaCompare/SchemaCompareSaveScmpOperation.cs
@@ -6,6 +6,7 @@
 #nullable disable
 using Microsoft.SqlServer.Dac.Compare;
 using Microsoft.SqlTools.ServiceLayer.Connection;
+using Microsoft.SqlTools.ServiceLayer.DacFx;
 using Microsoft.SqlTools.ServiceLayer.SchemaCompare.Contracts;
 using Microsoft.SqlTools.ServiceLayer.TaskServices;
 using Microsoft.SqlTools.Utility;
@@ -87,7 +88,7 @@ namespace Microsoft.SqlTools.ServiceLayer.SchemaCompare
 
                 if (this.Parameters.DeploymentOptions != null)
                 {
-                    comparison.Options = SchemaCompareUtils.CreateSchemaCompareOptions(this.Parameters.DeploymentOptions);
+                    comparison.Options = DacFxUtils.CreateSchemaCompareOptions(this.Parameters.DeploymentOptions);
                 }
 
                 comparison.SaveToFile(this.Parameters.ScmpFilePath, true);
